Ermm.. bug fixes but many things are still broken
OK. Logic maintained for mortgaging and selling properties, to sell a property (or mortgage) you must sell any and all houses in that color set. Test lines written for buying and selling houses. We're having the same problem there that we are with mortgaging, its not breaking after something is bought or mortgaged in the way it works with selling. Auction was added if you can't buy it.